refactor: add contexto entre rotas
